requirements.txt
google
streamlit
Pillow
requests
numpy
joblib
flask
scikit-learn
